// libs/features/payment/src/lib/shared/models/adjustment-repay-setItems.ts
import { AdjustmentRepay } from './beneficiary-list';

/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class AdjustmentRepaySetItems {
  repayItems: AdjustmentRepay[];
  totalAmount: number;
  adjustmentRepayId?: number;
}